Final Version?
OK - so I had to finish it! :)
Added billions, trillions, quadrillions etc.
Refactored common code.
Implemented console app.

+++ NumberSpeller/Int32Extender.cs
@@ -7,7 +7,7 @@ namespace NumberSpeller
 {
     public static class Int32Extender
     {
-        public static string ToText(this int number)
+        public static string ToText(this long number)
         {
             NumberString numStr = new NumberString(number);
             return numStr.Text;
